Reboot always after setting virthost to enable br0
We are setting br0 but we are not activating this.
We need to restart the network, but I don't want to do this and loose
the connectivity when salt is setting up configurations.
I think the safest thing is just to force the reboot always, to activate
this network configuration.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>